# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Job Application Tracker API",
    "version": "2.0.0",
    "description": "REST API for managing job applications and companies. Authentication is provided through Google OAuth 2.0 using Passport. Application and company endpoints require an authenticated session."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://company-applications-api.onrender.com",
      "description": "Render production server"
    }
  ],
...
  "paths": {
...
    "/auth/google/callback": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Google OAuth callback",
        "description": "Handles the callback from Google after authentication. Successful authentication creates a user session.",
        "responses": {
          "302": {
            "description": "Authentication successful and user is redirected to profile"
          },
          "401": {
            "description": "Google authentication failed"
          }
        }
      }
    },
    "/auth/profile": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get authenticated user profile",
        "description": "Returns the profile information of the currently authenticated Google user.",
        "responses": {
          "200": {
            "description": "Authenticated user profile returned successfully"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ]
      }
    },
...
    "/auth/logout": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logout current user",
        "description": "Logs out the authenticated user and destroys the current session.",
        "responses": {
          "200": {
            "description": "Logged out successfully"
          },
          "500": {
            "description": "Error while logging out"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie created after successful Google OAuth authentication."
      }
    },
...
  }
}